pyproject.toml
[build-system]
requires = ["setuptools>=69.1.1", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dMock"
version = "0.0.2"
description = "lightweight tool for mocking REST API dynamically"
requires-python = ">=3.11"
readme = "Readme.md"

dependencies = [
    "tortoise-orm>=0.20.0",
    "requests>=2.31.0",
    "aerich >= 0.7.2",
    "Hypercorn>=0.16.0",
    "toml>=0.10.2",
    "pytest-asyncio >= 0.23.6",
    "python-dotenv >= 1.0.1",
    "async-lru >= 2.0.4",
    "aiofiles >= 23.2.1", ]

authors = [
    { name = "Oleksii Ostapov" },
    { name = "Oleh Pleskach" }

]

classifiers = [
    "Topic :: Software Development :: Testing",
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
]

[project.urls]
Homepage = "https://infopulse.com"
Documentation = "https://github.com/infopulse/dMock"
Code = "https://github.com/infopulse/dMock"

[project.scripts]
dmock = "dmock.entry:main"

[tool.setuptools.packages.find]
exclude = [".github", "tests", "design", "dmock/models/migrations"]  # TODO check if migrations should be excluded

[tool.aerich]
tortoise_orm = "dmock.settings.DB_CONFIG"
location = "dmock/models/migrations"
src_folder = "./."
